backend/requirement.txt
# Core Dependencies
langchain==0.1.16
langchain-community==0.0.34
langchain-openai==0.1.3
langgraph==0.0.40

# Vector Store
chromadb==0.4.24
# 或使用 qdrant-client==1.8.0

# PDF Processing
pymupdf==1.23.26
pypdf==4.0.1
unstructured==0.12.4

# DOCX Processing
python-docx==1.1.0

# OCR Processing
pytesseract==0.3.10
Pillow==10.2.0

# Embeddings
sentence-transformers==2.5.1
torch==2.2.1

# API
fastapi==0.110.0
uvicorn[standard]==0.27.1
pydantic==2.6.3
python-multipart==0.0.9

# LLM Clients
openai==1.14.0
anthropic==0.21.3
dashscope==1.14.0  # 阿里通义千问

# Authentication
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0
PyJWT==2.8.0

# Utils
python-dotenv==1.0.1
tiktoken==0.6.0
redis==5.0.3
loguru==0.7.2
tenacity==8.2.3

# Testing
pytest==8.1.1
pytest-asyncio==0.23.5
httpx==0.27.0

# Database
sqlalchemy==2.0.28
alembic==1.13.1

# Additional dependencies for the complete system
requests==2.31.0
aiofiles==23.2.1
jinja2==3.1.2